refactor: LoginHandler 구조 추가
